chore: move detectResources back to different platform
